requirements.txt
requests
python-dotenv
flask
flask-cors
gunicorn
deep-translator
